댓글에 동영상 2개까지 가능 추가

+++ src/main/webapp/WEB-INF/jsp/article/detail.jsp
@@ -124,7 +124,11 @@ a {
 	<h2 class="con">댓글 작성</h2>
 
 	<script>
+		var ArticleWriteReplyForm__submitDone = false;
 		function ArticleWriteReplyForm__submit(form) {
+			if ( ArticleWriteReplyForm__submitDone ) {
+				alert('처리중입니다.');
+			}
 			form.body.value = form.body.value.trim();
 			if (form.body.value.length == 0) {
 				alert('댓글을 입력해주세요.');
@@ -132,10 +136,19 @@ a {
 				return;
 			}
 
+			ArticleWriteReplyForm__submitDone = true;
+
 			var startUploadFiles = function(onSuccess) {
-				var fileUploadFormData = new FormData(form);
+				if ( form.file__reply__0__common__attachment__1.value.length == 0 && form.file__reply__0__common__attachment__2.value.length == 0 ) {
+					onSuccess();
+					return;
+				}
+
+				var fileUploadFormData = new FormData(form); 
+				
 				fileUploadFormData.delete("relTypeCode");
 				fileUploadFormData.delete("relId");
+				fileUploadFormData.delete("body");
 
 				$.ajax({
 					url : './../file/doUploadAjax',
@@ -149,6 +162,7 @@ a {
 			}
 
 			var startWriteReply = function(fileIdsStr, onSuccess) {
+
 				$.ajax({
 					url : './../reply/doWriteReplyAjax',
 					data : {
@@ -164,9 +178,22 @@ a {
 			};
 
 			startUploadFiles(function(data) {
-				var idsStr = data.body.fileIdsStr;
+				
+				var idsStr = '';
+				if ( data && data.body && data.body.fileIdsStr ) {
+					idsStr = data.body.fileIdsStr;
+				}
+
 				startWriteReply(idsStr, function(data) {
+					
+					if ( data.msg ) {
+						alert(data.msg);
+					}
+					
 					form.body.value = '';
+					form.file__reply__0__common__attachment__1.value = '';
+					form.file__reply__0__common__attachment__2.value = '';
+					ArticleWriteReplyForm__submitDone = false;
 				});
 			});
 		}
@@ -192,8 +219,17 @@ a {
 					<th>첨부1 비디오</th>
 					<td>
 						<div class="form-control-box">
-							<input type="file" accept="video/*" capture
+							<input type="file" accept="video/*"
 								name="file__reply__0__common__attachment__1">
+						</div>
+					</td>
+				</tr>
+				<tr>
+					<th>첨부2 비디오</th>
+					<td>
+						<div class="form-control-box">
+							<input type="file" accept="video/*"
+								name="file__reply__0__common__attachment__2">
 						</div>
 					</td>
 				</tr>
@@ -303,8 +339,7 @@ a {
 		}, function(data) {
 			if (data.resultCode && data.resultCode.substr(0, 2) == 'S-') {
 				// 성공시에는 기존에 그려진 내용을 수정해야 한다.!!
-				var $tr = $('.reply-list-box tbody > tr[data-id="' + id
-						+ '"] .reply-body');
+				var $tr = $('.reply-list-box tbody > tr[data-id="' + id + '"] .reply-body');
 				$tr.empty().append(body);
 			}
 
@@ -378,9 +413,13 @@ a {
 		html += '<td>' + reply.extra.writer + '</td>';
 		html += '<td>';
 		html += '<div class="reply-body">' + reply.body + '</div>';
-		if (reply.extra.file__common__attachment__1) {
-            var file = reply.extra.file__common__attachment__1;
-            html += '<video controls src="http://localhost:8085/usr/file/streamVideo?id=' + file.id + '">video not supported</video>';
+
+		if ( reply.extra.file__common__attachment ) {
+			for ( var no in reply.extra.file__common__attachment ) {
+				var file = reply.extra.file__common__attachment[no];
+	            html += '<div class="video-box"><video controls src="/usr/file/streamVideo?id=' + file.id + '">video not supported</video></div>';				
+			}
+            
         }
 		
 		html += '</td>';
